fix(model): midtrans model

+++ database/migrations/2025_06_06_162544_create_subscription_plans_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
             $table->string('name');
             $table->string('slug')->unique();
             $table->text('description');
-            $table->decimal('price', 5, 2);
+            $table->integer('price');
             $table->enum('duration', ['monthly', 'yearly']);
 
             $table->integer('max_quizzes')->nullable();
